Mencari makan setiap hari

+++ src/CraftingTable.cpp
@@ -2,11 +2,8 @@
 
 CraftingTable::CraftingTable() : Inventory(3, 3, 'C') {}
 
-CraftingTable::CraftingTable(const CraftingTable& ct)
-    : Inventory(ct) {}
-Inventory* CraftingTable::clone(){
-  return new CraftingTable(*this);
-}
+CraftingTable::CraftingTable(const CraftingTable& ct) : Inventory(ct) {}
+Inventory* CraftingTable::clone() { return new CraftingTable(*this); }
 bool CraftingTable::operator==(const CraftingTable& ct) {
   CraftingTable cp(ct);
 
@@ -18,19 +15,19 @@ bool CraftingTable::operator==(const CraftingTable& ct) {
       }
     }
   }
-  
+
   return true;
 }
 
-void CraftingTable::craft(){
+void CraftingTable::craft() {
   CraftingTable* tempC = new CraftingTable(*this);
-  try{
-    for(int i=0; i<this->mxRow; i++){
-      for(int j=0; j<this->mxCol; j++){
-        this->deleteItem({i, j}, 1);
+  try {
+    for (int i = 0; i < this->mxRow; i++) {
+      for (int j = 0; j < this->mxCol; j++) {
+        if (!this->container[i][j].empty()) this->deleteItem({i, j}, 1);
       }
     }
-  }catch(BaseException* e){
+  } catch (BaseException* e) {
     this->operator=(*tempC);
     throw e;
   }
